+++ src/novelystlib/project_view.py
@@ -20,7 +20,12 @@ class ProjectView(ElementView):
         Extends the superclass constructor.
         """
         super(). __init__(ui, element)
-        self._renChapters = tk.BooleanVar(value=element.kwVar['Field_RenumberChapters'])
+        self._authorName = tk.StringVar(value=element.authorName)
+        self._authorNameEntry = LabelEntry(self._valuesFrame, text='Author', textvariable=self._authorName, lblWidth=20)
+        self._authorNameEntry.pack(anchor='w', pady=2)
+
+        renChapters = element.kwVar['Field_RenumberChapters'] == '1'
+        self._renChapters = tk.BooleanVar(value=renChapters)
         self._renChaptersCheckbox = ttk.Checkbutton(self._valuesFrame, text='Auto number chapters when refreshing the tree',
                                          variable=self._renChapters, onvalue=True, offvalue=False)
         self._renChaptersCheckbox.pack(anchor='w', pady=2)
@@ -33,17 +38,20 @@ class ProjectView(ElementView):
         self._chHdSuffixEntry = LabelEntry(self._valuesFrame, text='Chapter heading suffix', textvariable=self._chHdSuffix, lblWidth=20)
         self._chHdSuffixEntry.pack(anchor='w', pady=2)
 
-        self._romanChapters = tk.BooleanVar(value=element.kwVar['Field_RomanChapterNumbers'])
+        romanChapters = element.kwVar['Field_RomanChapterNumbers'] == '1'
+        self._romanChapters = tk.BooleanVar(value=romanChapters)
         self._romanChaptersCheckbox = ttk.Checkbutton(self._valuesFrame, text='Use Roman chapter numbers',
                                         variable=self._romanChapters, onvalue=True, offvalue=False)
         self._romanChaptersCheckbox.pack(anchor='w', pady=2)
 
-        self._renWithinParts = tk.BooleanVar(value=element.kwVar['Field_RenumberWithinParts'])
+        renWithinParts = element.kwVar['Field_RenumberWithinParts'] == '1'
+        self._renWithinParts = tk.BooleanVar(value=renWithinParts)
         self._renWithinPartsCheckbox = ttk.Checkbutton(self._valuesFrame, text='Reset chapter number when starting a new part',
                                         variable=self._renWithinParts, onvalue=True, offvalue=False)
         self._renWithinPartsCheckbox.pack(anchor='w', pady=2)
 
-        self._renParts = tk.BooleanVar(value=element.kwVar['Field_RenumberParts'])
+        renParts = element.kwVar['Field_RenumberParts'] == '1'
+        self._renParts = tk.BooleanVar(value=renParts)
         self._renPartsCheckbox = ttk.Checkbutton(self._valuesFrame, text='Auto number parts when refreshing the tree',
                                         variable=self._renParts, onvalue=True, offvalue=False)
         self._renPartsCheckbox.pack(anchor='w', pady=2)
@@ -56,7 +64,8 @@ class ProjectView(ElementView):
         self._ptHdSuffixEntry = LabelEntry(self._valuesFrame, text='Part heading suffix', textvariable=self._ptHdSuffix, lblWidth=20)
         self._ptHdSuffixEntry.pack(anchor='w', pady=2)
 
-        self._romanParts = tk.BooleanVar(value=element.kwVar['Field_RomanPartNumbers'])
+        romanParts = element.kwVar['Field_RomanPartNumbers'] == 1
+        self._romanParts = tk.BooleanVar(value=romanParts)
         self._romanPartsCheckbox = ttk.Checkbutton(self._valuesFrame, text='Use Roman part numbers',
                                         variable=self._romanParts, onvalue=True, offvalue=False)
         self._romanPartsCheckbox.pack(anchor='w', pady=2)
@@ -66,28 +75,28 @@ class ProjectView(ElementView):
         
         Extends the superclass method.
         """
-
-        def update_field(tkValue, fieldname):
-            """Update a custom field if changed.
-            
-            Positional arguments:
-                tkValue -- widget variable holding a value that is not None.
-                fieldname -- keyword of a custom field whose value might be None.
-            """
-            entry = tkValue.get()
-            if self._element.kwVar[fieldname] or entry:
-                if self._element.kwVar[fieldname] != entry:
-                    self._element.kwVar[fieldname] = entry
-                    ui.isModified = True
-
-        update_field(self._renChapters, 'Field_RenumberChapters')
-        update_field(self._chHdPrefix, 'Field_ChapterHeadingPrefix')
-        update_field(self._chHdSuffix, 'Field_ChapterHeadingSuffix')
-        update_field(self._romanChapters, 'Field_RomanChapterNumbers')
-        update_field(self._renWithinParts, 'Field_RenumberWithinParts')
-        update_field(self._renParts, 'Field_RenumberParts')
-        update_field(self._ptHdPrefix, 'Field_PartHeadingPrefix')
-        update_field(self._ptHdSuffix, 'Field_PartHeadingSuffix')
-        update_field(self._romanParts, 'Field_RomanPartNumbers')
+        authorName = self._authorName.get()
+        if authorName or self._element.authorName:
+            if self._element.authorName != authorName:
+                self._element.authorName = authorName.strip()
+                ui.isModified = True
+        if self._update_field_bool(self._renChapters, 'Field_RenumberChapters'):
+                ui.isModified = True
+        if self._update_field_str(self._chHdPrefix, 'Field_ChapterHeadingPrefix'):
+                ui.isModified = True
+        if self._update_field_str(self._chHdSuffix, 'Field_ChapterHeadingSuffix'):
+                ui.isModified = True
+        if self._update_field_bool(self._romanChapters, 'Field_RomanChapterNumbers'):
+                ui.isModified = True
+        if self._update_field_bool(self._renWithinParts, 'Field_RenumberWithinParts'):
+                ui.isModified = True
+        if self._update_field_bool(self._renParts, 'Field_RenumberParts'):
+                ui.isModified = True
+        if self._update_field_str(self._ptHdPrefix, 'Field_PartHeadingPrefix'):
+                ui.isModified = True
+        if self._update_field_str(self._ptHdSuffix, 'Field_PartHeadingSuffix'):
+                ui.isModified = True
+        if self._update_field_bool(self._romanParts, 'Field_RomanPartNumbers'):
+                ui.isModified = True
         super().apply_changes(ui)